// Admin/src/app/interfaces/rider.ts
export interface Rider {
    Loc_lat: number;
    Loc_lng: number;
    Rider_id: string;
    Rider_name: string;
    Rider_email: string;
    Rider_phone: number;
    Rider_imgUrl: string;
    photoUrl?: string;
    Rider_rating: number;
    Des_lat: number;
    Des_lng: number;
    Rider_Location:  string;
    Rider_Destination: string;
    countDown: number;
    cancel: boolean;
    price: number;
    cash: boolean;
    time: any;
    notificationID: any
}